extract input code to separate file
